Adding data retrieving
